Metrics mostly complete
