+++ src/MUST_analysis_tools/MUST_CP_long_analysis.py
@@ -11,6 +11,14 @@ from matplotlib.image import imread
 import os
 import pyproj
 
+# ## Notes about this file:
+
+# # Methodology:
+# Most of the data analysis is done with resampled data rather than raw. The vehicle GPS used in the original UW testing
+#   in Washington was the novatel PwrPak6D-E2 from the white pacifica, ~50hz output frequency and ~10-30cm accuracy. The
+#   output from the MUST sensor was ~12-15hz. Both data series were resampled to a consistent 50hz with the same main
+#   time series so they are directly comparable for the metrics.
+
 # GPS stationary noise accounting - Make speed zero for values below this number
 MUST_STATIONARY_NOISE = 0.1 #meter/sec
 GPS_STATIONARY_NOISE = 0.1 #meter/sec
@@ -69,111 +77,6 @@ def compute_distance_error_haversine(lat1_orig, lon1_orig, lat2, lon2, offset):
     return distance
 
 
-def resample_df(df_original, period_ms):
-    columns = ['latitude', 'longitude', 'speed', 'heading']
-    start_time = df_original['epoch_time'][0]
-    end_time = df_original['epoch_time'][len(df_original)-1]
-    start_time_smoothed = period_ms * np.ceil(start_time / period_ms)
-    end_time_smoothed = period_ms * np.floor(end_time / period_ms)
-    df_resampled = pd.DataFrame()
-    df_resampled['epoch_time'] = np.arange(start_time_smoothed, end_time_smoothed, period_ms)
-    for column in columns:
-        df_resampled[column] = resample_series(df_resampled['epoch_time'].to_numpy(), df_original['epoch_time'].to_numpy(), df_original[column].to_numpy())
-    return df_resampled
-
-
-# from https://stackoverflow.com/questions/2566412/find-nearest-value-in-numpy-array
-def find_nearest_index(array,value):
-    idx = np.searchsorted(array, value, side="left")
-    if idx > 0 and (idx == len(array) or math.fabs(value - array[idx-1]) < math.fabs(value - array[idx])):
-        return idx-1
-    else:
-        return idx
-
-
-# Assumes that main_time spans series_time, slightly further on both sides
-def resample_series(main_time, series_time, series_data):
-    overlap_length = len(main_time)
-
-    series_data_main_times = np.zeros(overlap_length)
-    for i in range(len(main_time)):
-        t = main_time[i]
-        if np.min(np.abs(series_time - t)) < 0.001:
-            series_index = np.argmin(np.abs(series_time - t))
-            series_data_main_times[i] = series_data[series_index]
-            continue
-        delta = series_time - t
-        delta[delta < 0] = np.inf
-        i_before = max(np.argmin(delta) - 1, 0)
-        slope = ((series_data[i_before+1] - series_data[i_before]) / (series_time[i_before+1] - series_time[i_before]))
-        time_delta_before = series_time[i_before] - t
-        series_data_main_times[i] = series_data[i_before] + slope * time_delta_before
-    return series_data_main_times
-
-
-# Assumes that df1 spans df2 before and after the 'optimal' time alignment.
-# Assumes x1/y1 and x2/y2
-# Steps df2 over each piece of df1 computing the minimum distance error for each time shift
-# returns the index of that optimal time shift
-def find_optimal_time_shift(lat1, lon1, lat2, lon2):
-    overlap_region = len(lat1) - len(lat2)
-    dist_errors = np.zeros(overlap_region)
-    for i in range(overlap_region):
-        dist_errors[i] = compute_distance_error_haversine(lat1, lon1, lat2, lon2, i)
-    min_error_index = np.argmin(dist_errors)
-    return min_error_index
-
-
-def find_sub_tracks(time, lat, lon):
-    break_time = 5.0 # seconds
-    break_dist = 5 # meters
-    break_indices = []
-    for i in range(1, len(time)):
-        if haversine_distance(lat[i-1], lon[i-1], lat[i], lon[i]) > break_dist or time[i] - time[i-1] > break_time:
-            break_indices.append(i)
-    break_indices = [0] + break_indices + [len(time)]
-    return break_indices
-
-
-def select_sub_track_user_input(must_data, gps_data, test_name, intersection_image_path):
-
-    # Set up the figure and axes
-    fig, ax = plt.subplots(figsize=(10, 10))
-
-    # Create a Basemap instance
-    intersection_image = imread(intersection_image_path)
-    # Set the bounds of the map (in lat/lon coordinates)
-    lower_left_longitude = -122.143605  # Lower-left corner longitude
-    lower_left_latitude = 47.627545  # Lower-left corner latitude
-    upper_right_longitude = -122.142310  # Upper-right corner longitude
-    upper_right_latitude = 47.628340  # Upper-right corner latitude
-    map = Basemap(projection='merc', llcrnrlat=lower_left_latitude, urcrnrlat=upper_right_latitude,
-                llcrnrlon=lower_left_longitude, urcrnrlon=upper_right_longitude, resolution='i', ax=ax)
-
-    # Show the image as the background
-    map.imshow(intersection_image, origin='upper')
-    gps_x, gps_y = map(gps_data['longitude'], gps_data['latitude'])
-    map.plot(gps_x, gps_y, marker='o', markersize=5, label=f'GPS track')
-
-    break_indices = find_sub_tracks(must_data['epoch_time'].to_numpy(), must_data['latitude'].to_numpy(), must_data['longitude'].to_numpy())
-
-    for i in range(len(break_indices) - 1):
-        sub_track_indices = range(break_indices[i], break_indices[i + 1])
-        # Convert lat/lon to map projection coordinates
-        must_x, must_y = map(must_data['longitude'].iloc[sub_track_indices], must_data['latitude'].iloc[sub_track_indices])
-
-        # Plot the data points
-        map.plot(must_x, must_y, marker='o', markersize=5, label=f'Track {i}')
-
-    # Add labels and a legend
-    plt.xlabel('Longitude')
-    plt.ylabel('Latitude')
-    plt.legend()
-    plt.title(f'Vehicle tracks {test_name}')
-    plt.tight_layout()
-    plt.show()
-
-
 def load_metadata(test_name, test_log_fname, gps_folder):
     header = ['Test numbering', 'Test Case ID', 'Run #', 'Corrected?', 'Original',
               'TimeStamp', 'Description', 'Vehicle Speed', 'Vehicle Lane', 'Date', 'Pass/Fail',
@@ -222,54 +125,7 @@ def generate_plots(test_name, test_log, intersection_image_path, gps_folder, mus
     # Heading is North, positive West -> North, positive East
     must_data['heading'] = (must_data['heading']) % 360
 
-    gps_header = ['timestamp', 'latitude', 'longitude', 'altitude', 'heading', 'speed', 'latitude stdev', 'longitude stdev', 'altitude stdev', 'heading error', 'speed error']
-    gps_data = pd.read_csv(f'{gps_folder}\\{gps_filename}', names=gps_header, skiprows=1)
-    gps_data['epoch_time'] = gps_data['timestamp']
-
-    if 'track_index' not in locals() or track_index is None:
-        return
-        select_sub_track_user_input(must_data, gps_data, test_name, intersection_image_path)
-        return
-    # else:
-    #     return
-
-    break_indices = break_indices = find_sub_tracks(must_data['epoch_time'].to_numpy(),
-                                                    must_data['latitude'].to_numpy(), must_data['longitude'].to_numpy())
-    valid_data_indices = range(break_indices[track_index], break_indices[track_index + 1])
-    must_data = must_data.iloc[valid_data_indices[1:]].reset_index(drop=True)
-
-    # gps_data['datetime'] = pd.to_datetime(gps_data['epoch_time'], unit='s')
-    # gps_data = gps_data.set_index('datetime')
-    # gps_data.index = gps_data.index.to_series().dt.round('1ms')
-    # must_data['datetime'] = pd.to_datetime(must_data['epoch_time'], unit='s')
-    # must_data = must_data.set_index('datetime')
-    # must_data.index = must_data.index.to_series().dt.round('1ms')
-    # must_data = must_data.reindex(gps_data.index).interpolate(method='time')
-    gps_data_resampled = resample_df(gps_data, 0.02)
-    must_data_resampled = resample_df(must_data, 0.02)
-
-    # example_point = [47.6278859, -122.1431239]
-    # must_distances = [haversine_distance(example_point[0], example_point[1], must_data['latitude'][i], must_data['longitude'][i])
-    #              for i in range(len(must_data['latitude']))]
-    # must_min_index = np.argmin(must_distances)
-    # gps_distances = [haversine_distance(example_point[0], example_point[1], gps_data['latitude'][i], gps_data['longitude'][i])
-    #              for i in range(len(gps_data['latitude']))]
-    # gps_min_index = np.argmin(gps_distances)
-    # must_min_epoch = must_data['epoch_time'][must_min_index]
-    # gps_min_epoch = gps_data['epoch_time'][gps_min_index]
-    # gps_time_offset = must_min_epoch - gps_min_epoch
-    # gps_time_offset = -117.5
-    time_offset_index = find_optimal_time_shift(gps_data_resampled['latitude'].to_numpy(), gps_data_resampled['longitude'].to_numpy(),
-                                                must_data_resampled['latitude'].to_numpy(), must_data_resampled['longitude'].to_numpy())
-    time_offset = gps_data_resampled['epoch_time'][time_offset_index] - gps_data_resampled['epoch_time'][0]
-    gps_data_matched = gps_data_resampled
-    must_data_matched = must_data_resampled
-    must_data_matched['epoch_time'] = must_data_matched['epoch_time'] + time_offset
-    gps_data_matched = gps_data_matched[time_offset_index-1:time_offset_index-1 + len(must_data_matched)].reset_index(drop=True)
-
-    print(f'time offset for test {test_name}: {time_offset}, static: {-gps_data['epoch_time'][0] + must_data['epoch_time'][0] + 117.5}')
-    gps_data['sim time'] = gps_data['epoch_time'] - gps_data['epoch_time'][0]#  + (gps_data['epoch_time'][0] - must_data['epoch_time'][0] - 117.5)
-    must_data['sim time'] = must_data['epoch_time'] - must_data['epoch_time'][0] + time_offset
+    must_data['sim time'] = must_data['epoch_time'] - must_data['epoch_time'][0]
     # Set up the figure and axes
     fig = plt.figure(figsize=(20, 12), dpi=100)
     gs = gridspec.GridSpec(2, 3, figure=fig)
@@ -297,6 +153,9 @@ def generate_plots(test_name, test_log, intersection_image_path, gps_folder, mus
                                                     must_data['longitude'].iloc[i + 1])
                                 for i in range(len(must_data) - 1)]
     must_data['heading_interpolated'] = (np.convolve(must_data['heading_interpolated'].to_numpy(), np.ones(N) / N, mode='same') % 360)
+
+    must_data['heading_unwrapped'] = np.unwrap(must_data['heading'], period=360, discont=270)
+    gps_data['heading_unwrapped'] = np.unwrap(gps_data['heading'], period=360, discont=270)
 
     shadow_data = gps_data.copy()
     vehicle_height = 0.5
@@ -359,11 +218,9 @@ def generate_plots(test_name, test_log, intersection_image_path, gps_folder, mus
     ax2 = fig.add_subplot(gs[0, 2])
     ax3 = fig.add_subplot(gs[1, 2])
     ax2.plot(must_data['sim time'], must_data['speed'], label='MUST speed')
-    ax2.plot(gps_data['sim time'], gps_data['speed'], label='GPS speed')
     ax2.plot(must_data['sim time'], must_data['speed_interpolated'], label='MUST speed (interpolated)')
     ax2.legend()
-    ax3.plot(must_data['sim time'], must_data['heading'], label='MUST heading')
-    ax3.plot(gps_data['sim time'], gps_data['heading'], label='GPS heading')
+    ax3.plot(must_data['sim time'], must_data['heading_unwrapped'], label='MUST heading')
     ax3.plot(must_data['sim time'], must_data['heading_interpolated'], label='MUST heading (interpolated)')
 
     ax2.set_title('speed vs. time')
@@ -384,7 +241,7 @@ def generate_plots(test_name, test_log, intersection_image_path, gps_folder, mus
     # plt.show()
     # plt.clf()
 
-    # Metric 1: position accuracy (90% <30cm error)
+    # Metric 6: confidence score >90% for >90% of vehicle tracks
     distance_errors = np.array([haversine_distance(must_data_matched['latitude'].iloc[i],
                                           must_data_matched['longitude'].iloc[i],
                                           gps_data_matched['latitude'].iloc[i],
@@ -393,19 +250,28 @@ def generate_plots(test_name, test_log, intersection_image_path, gps_folder, mus
     distance_mean = np.mean(distance_errors)
     distance_stdev = np.std(distance_errors)
     distances_pct_below_limit = 100 * len(distance_errors[distance_errors < 0.3]) / len(distance_errors)
-    print(f'Metric 1, distance error <30cm. Mean error: {distance_mean}, stdev: {distance_stdev}, percentage below 30cm: {distances_pct_below_limit}%')
+    # insert histogram. X axis confidence score, Y axis number of tracks
+    print(f'Metric 6, distance error <30cm. Mean error: {distance_mean}, stdev: {distance_stdev}, percentage below 30cm: {distances_pct_below_limit}%')
 
-    # Metric 2: speed accuracy (90% <3mph error)
-    speeds = np.array([haversine_distance(must_data_matched['latitude'].iloc[i],
+    # Metric 7: detection frequency (30hz +- 3hz)
+    recording_freqs = np.array([1 / (must_data['epoch_time'].iloc[i+1] - must_data['epoch_time'].iloc[i])
+                       for i in range(len(must_data) - 1)])
+    freq_mean = np.mean(recording_freqs)
+    freq_stdev = np.std(distance_errors)
+    freq_pct_below_limit = 100 * len(recording_freqs[recording_freqs > 30]) / len(recording_freqs)
+    # insert histogram. X axis frequency (0-30), Y axis percentage
+    print(f'Metric 7, distance error <30cm. Mean error: {distance_mean}, stdev: {distance_stdev}, percentage below 30cm: {distances_pct_below_limit}%')
+
+    # Metric 8: Object type staying consistent (90% object types the same)
+    distance_errors = np.array([haversine_distance(must_data_matched['latitude'].iloc[i],
                                           must_data_matched['longitude'].iloc[i],
                                           gps_data_matched['latitude'].iloc[i],
                                           gps_data_matched['longitude'].iloc[i])
                        for i in range(len(must_data_matched) - 1)])
-    distance_mean = np.mean(distances)
-    distance_stdev = np.std(distances)
-    distances_pct_below_limit = 100 * len(distances[distances < 0.3]) / len(distances)
-    print(f'Metric 2, speed error <3 mph. Mean error: {distance_mean}, stdev: {distance_stdev}, percentage below 3 mph: {distances_pct_below_limit}%')
-
+    distance_mean = np.mean(distance_errors)
+    distance_stdev = np.std(distance_errors)
+    distances_pct_below_limit = 100 * len(distance_errors[distance_errors < 0.3]) / len(distance_errors)
+    print(f'Metric 8, distance error <30cm. Mean error: {distance_mean}, stdev: {distance_stdev}, percentage below 30cm: {distances_pct_below_limit}%')
 
 
 def main(args):
